chore: 1-click refactor for permission group service
