+++ app/server/appsmith-server/src/main/java/com/appsmith/server/services/ce_compatible/PermissionGroupServiceCECompatibleImpl.java
@@ -1,9 +1,12 @@
-package com.appsmith.server.services;
+package com.appsmith.server.services.ce_compatible;
 
 import com.appsmith.server.repositories.ConfigRepository;
 import com.appsmith.server.repositories.PermissionGroupRepository;
 import com.appsmith.server.repositories.UserRepository;
-import com.appsmith.server.services.ce_compatible.PermissionGroupServiceCECompatibleImpl;
+import com.appsmith.server.services.AnalyticsService;
+import com.appsmith.server.services.SessionUserService;
+import com.appsmith.server.services.TenantService;
+import com.appsmith.server.services.ce.PermissionGroupServiceCEImpl;
 import com.appsmith.server.solutions.PermissionGroupPermission;
 import com.appsmith.server.solutions.PolicySolution;
 import jakarta.validation.Validator;
@@ -13,10 +16,9 @@ import org.springframework.stereotype.Service;
 import reactor.core.scheduler.Scheduler;
 
 @Service
-public class PermissionGroupServiceImpl extends PermissionGroupServiceCECompatibleImpl
-        implements PermissionGroupService {
-
-    public PermissionGroupServiceImpl(
+public class PermissionGroupServiceCECompatibleImpl extends PermissionGroupServiceCEImpl
+        implements PermissionGroupServiceCECompatible {
+    public PermissionGroupServiceCECompatibleImpl(
             Scheduler scheduler,
             Validator validator,
             MongoConverter mongoConverter,
@@ -29,7 +31,6 @@ public class PermissionGroupServiceImpl extends PermissionGroupServiceCECompatib
             PolicySolution policySolution,
             ConfigRepository configRepository,
             PermissionGroupPermission permissionGroupPermission) {
-
         super(
                 scheduler,
                 validator,
